Makes core exported library more compact
